+++ app/Http/Controllers/TestCategoryController.php
@@ -6,6 +6,7 @@ use App\Models\Category;
 use App\Models\Question;
 use App\Models\TestCategory;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Session;
 
 class TestCategoryController extends Controller
@@ -13,6 +14,10 @@ class TestCategoryController extends Controller
 
     public function create($test_id)
     {
+        if(Auth::user() == null || !Auth::user()->hasRole('profesor'))
+        {
+            return redirect()->route('home');
+        }
         $categories = Category::all()->pluck('name', 'id');
 
         return view('test_category.create', compact('test_id', 'categories'));
@@ -20,11 +25,19 @@ class TestCategoryController extends Controller
 
     public function show(Category $category, Question $question)
     {
+        if(Auth::user() == null || !Auth::user()->hasRole('profesor'))
+        {
+            return redirect()->route('home');
+        }
         return view('categories.show', compact( 'question'));
     }
 
     public function edit($test_id, $category_id)
     {
+        if(Auth::user() == null || !Auth::user()->hasRole('profesor'))
+        {
+            return redirect()->route('home');
+        }
 
         $test_category = TestCategory::all()->where('test_id', '=', $test_id)->where('category_id', '=', $category_id)->first();
         $categories = Category::all();
@@ -34,7 +47,10 @@ class TestCategoryController extends Controller
 
     public function store(Request $request, $test_id)
     {
-
+        if(Auth::user() == null || !Auth::user()->hasRole('profesor'))
+        {
+            return redirect()->route('home');
+        }
         $test_category = new TestCategory();
 
         $test_category->category_id = $request->category_id;
@@ -49,6 +65,12 @@ class TestCategoryController extends Controller
 
     public function update(Request $request, $category_id, $test_id)
     {
+
+        if(Auth::user() == null || !Auth::user()->hasRole('profesor'))
+        {
+            return redirect()->route('home');
+        }
+
         $test_category = TestCategory::all()->where('category_id', '=', $category_id)->where('test_id', '=', $test_id)->first();
         $test_category->number_of_questions = $request->number_of_questions;
         $test_category->update();
@@ -60,6 +82,11 @@ class TestCategoryController extends Controller
 
     public function destroy( $test_id, $category_id)
     {
+
+        if(Auth::user() == null || !Auth::user()->hasRole('profesor'))
+        {
+            return redirect()->route('home');
+        }
         $test_category = TestCategory::all()->where('test_id', '=', $test_id)->where('category_id', '=', $category_id)->first();
         $test_category->delete();
         Session::flash('delete-message', 'Question deleted successfully');
